First version of map lookup
Uses operator[]. Code is easy to write and read, but behavior is
somewhat surprising. If user enters a color name that isn't in the map,
then the code prints "(0,0,0)". Because this is a vaid return value
(it's the color code for black), there's no reliable way to detect this
error condition.