Attempt to refactor map access into a function
In this simple example code, there's not a very good reason to do this
refactoring. But let's do it anyway.

It should be possible to write a function that reads a
`const std::map<...>&`. The current code doesn't compile. The problem is
that `std::map::operator[]` is a non-const operation. If the request key
doesn't exist, `operator[]` inserts a new mapping (with the value
default-constructed).  That's why `colors[name] = color` works in the
first half of main, but it's also why getColor fails.