Use map.at() instead of map.operator[]
map.at() is a const function (or at least it has a const overload). Now
the code compiles and runs. But if the user enters an unknown color
name, the code crashes (technically: map.at() throws an exception; we
don't catch the exception; C++ calls std::terminate() to end our
program).